split 'build-hook' into 'utils'
so we can build several utilities sharing some code.
